+++ src/ts/elk-sprotty-bridge/index.d.ts
@@ -1,0 +1,3 @@
+import { ElkGraphJsonToSprotty } from "./elkgraph-to-sprotty";
+import { createContainer } from './sprotty-config';
+export { createContainer, ElkGraphJsonToSprotty, };